pyproject.toml
[project]
name = "projeto-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.189.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "langchain>=1.2.8",
    "langchain-core>=1.2.8",
    "langchain-openai>=1.1.7",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
